Add files via upload
Weekend Challenge: Fundamentals
Create a fill-in-the blank quiz game. Ask the users's name, then refer to the user by name as you ask him/her a series of 
questions that you have stored in an array. Use the prompt() function to get each input from the user and compare it to the answer you expected. When the user enters "Q" (for quit), or perhaps when the user hits [Cancel], exit the game and prints 
the statistics of the game to the console: user name, number of questions answered and questions correct.